refactor: split settings into organization, user, and invoice design pages
